Fixed #1813

+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/gadgets/SolarHelmet.java
@@ -10,7 +10,7 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
 public class SolarHelmet extends SlimefunItem {
 
-    private final ItemSetting<Float> chargeSetting = new ItemSetting<>("charge-amount", 0.1F);
+    private final ItemSetting<Double> chargeSetting = new ItemSetting<>("charge-amount", 0.1);
 
     public SolarHelmet(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe, null);
@@ -18,7 +18,7 @@ public class SolarHelmet extends SlimefunItem {
         addItemSetting(chargeSetting);
     }
 
-    public float getChargeAmount() {
+    public double getChargeAmount() {
         return chargeSetting.getValue();
     }
